+++ Platipus.Wallet.Api/Application/Services/Wallet/IWalletService.cs
@@ -35,7 +35,7 @@ public interface IWalletService
         string transactionId,
         string? roundId = null,
         bool searchByUsername = false,
-        int? amount = null,
+        decimal? amount = null,
         string? clientId = null,
         CancellationToken cancellationToken = default);
