+++ Platipus.Wallet.Api/Application/Services/Wallet/WalletService.cs
@@ -225,7 +225,7 @@ public sealed class WalletService : IWalletService
         string transactionId,
         string? roundId = null,
         bool searchByUsername = false,
-        int? amount = null,
+        decimal? amount = null,
         string? clientId = null,
         CancellationToken cancellationToken = default)
     {
